Video delete from S3 (#14)
* database schema init and updated migration and model files (#1)

* database schema init and updated migration and model files

* updated table models as per the feedback and clean the code

* removing seed file

* added createUser query and schema on it and started with invitation table  (#2)

* added createUser query and scema on it and removed seed file from knex.

* more updates on invitation

* invite flow and validation with create user (#3)

* More updates on Query update around the add and fetch project and videos functionality and added remove user and workspace functionality with required validations (#4)

* invite flow and validation with create user

* api updated on remove workspace , users  and updated project and videos queries based on user id and workspaceid of the user.

* minute fix for invite flow

* updated project and videos mutation and queries

* updating migration file for foreign key issue (#5)

* invite flow and validation with create user

* api updated on remove workspace , users  and updated project and videos queries based on user id and workspaceid of the user.

* minute fix for invite flow

* updated project and videos mutation and queries

* updating migration file for foreign key issue.

* Refactor database schema and updated create user flow in case of invite flow (#6)

* invite flow and validation with create user

* api updated on remove workspace , users  and updated project and videos queries based on user id and workspaceid of the user.

* minute fix for invite flow

* updated project and videos mutation and queries

* updating migration file for foreign key issue.

* updated model and database schema and updated create user flow in case of invite flow

* Updates around userInfo flow and invitation id (#7)

* invite flow and validation with create user

* api updated on remove workspace , users  and updated project and videos queries based on user id and workspaceid of the user.

* minute fix for invite flow

* updated project and videos mutation and queries

* updating migration file for foreign key issue.

* updated model and database schema and updated create user flow in case of invite flow

* invitation and getInfo Query changes

* added uuid for invite

* updates on userinfo schema

* sprint3 - init: updated database schema and model for sprint 3 work (#9)

* invite flow and validation with create user

* api updated on remove workspace , users  and updated project and videos queries based on user id and workspaceid of the user.

* minute fix for invite flow

* updated project and videos mutation and queries

* updating migration file for foreign key issue.

* updated model and database schema and updated create user flow in case of invite flow

* invitation and getInfo Query changes

* added uuid for invite

* updates on userinfo schema

* updated database schema and model for sprint 3 work

* topic init and updated project and videos related files

* more updates on topic and video api

* Data loader init for topics and videos in project call. (#10)

* Staging (#8)

* database schema init and updated migration and model files (#1)

* database schema init and updated migration and model files

* updated table models as per the feedback and clean the code

* removing seed file

* added createUser query and schema on it and started with invitation table  (#2)

* added createUser query and scema on it and removed seed file from knex.

* more updates on invitation

* invite flow and validation with create user (#3)

* More updates on Query update around the add and fetch project and videos functionality and added remove user and workspace functionality with required validations (#4)

* invite flow and validation with create user

* api updated on remove workspace , users  and updated project and videos queries based on user id and workspaceid of the user.

* minute fix for invite flow

* updated project and videos mutation and queries

* updating migration file for foreign key issue (#5)

* invite flow and validation with create user

* api updated on remove workspace , users  and updated project and videos queries based on user id and workspaceid of the user.

* minute fix for invite flow

* updated project and videos mutation and queries

* updating migration file for foreign key issue.

* Refactor database schema and updated create user flow in case of invite flow (#6)

* invite flow and validation with create user

* api updated on remove workspace , users  and updated project and videos queries based on user id and workspaceid of the user.

* minute fix for invite flow

* updated project and videos mutation and queries

* updating migration file for foreign key issue.

* updated model and database schema and updated create user flow in case of invite flow

* Updates around userInfo flow and invitation id (#7)

* invite flow and validation with create user

* api updated on remove workspace , users  and updated project and videos queries based on user id and workspaceid of the user.

* minute fix for invite flow

* updated project and videos mutation and queries

* updating migration file for foreign key issue.

* updated model and database schema and updated create user flow in case of invite flow

* invitation and getInfo Query changes

* added uuid for invite

* updates on userinfo schema


* data loader init for topics and videos in project call.

* updated dataloader queries


* video delete from aws s3 bucket (#13)

* addded s3Utils in order to make it singleton

+++ src/services/video.service.js
@@ -1,6 +1,7 @@
 import BaseService from './base.service';
 import Video from '../models/video.model';
 import { transaction } from 'objection';
+import S3UtilService from '../utils/s3Utils';
 
 class VideoService extends BaseService {
   constructor() {
@@ -40,22 +41,39 @@ class VideoService extends BaseService {
     }
   }
 
+  getVideoObject(videos) {
+    const videoObjArray = [];
+    videos.forEach((video) => {
+      videoObjArray.push({ Key: video.video });
+    });
+    return videoObjArray;
+  }
+
   async deleteVideo(id) {
     const video = await this.findById(id);
-
+    await S3UtilService.deleteObjects(this.getVideoObject([video]));
     await Video.query().deleteById(id);
-
     return video;
   }
 
   async deleteVideoByProjectId(projectId) {
-    await Video.query().delete().where('projectId', projectId);
-    return true;
+    const videos = await Video.query().where('projectId', projectId);
+    if (videos.length) {
+      await S3UtilService.deleteObjects(this.getVideoObject(videos));
+      const result = await Video.query().delete().where('projectId', projectId);
+      return result;
+    }
+    return false;
   }
 
   async deleteVideoByTopicId(topicId) {
-    await Video.query().delete().where('topicId', topicId);
-    return true;
+    const videos = await Video.query().where('topicId', topicId);
+    if (videos.length) {
+      await S3UtilService.deleteObjects(this.getVideoObject(videos));
+      const result = await Video.query().delete().where('topicId', topicId);
+      return result;
+    }
+    return false;
   }
 
   async findByProject(projectId, orderBy = {}) {
@@ -71,7 +89,7 @@ class VideoService extends BaseService {
   }
 
   async findByTopics(topicIds, orderBy = {}) {
-    const { field = '', direction = 'asc' } = orderBy;
+    const { field = 'updatedAt', direction = 'desc' } = orderBy;
 
     let query = Video.query()
       .select('videos.*', 'users.firstName', 'users.lastName', 'users.email')
@@ -85,13 +103,14 @@ class VideoService extends BaseService {
     return query;
   }
 
-  async findByTopic(topicId, orderBy = {}) {
-    const { field = '', direction = 'asc' } = orderBy;
+  async findByTopic(projectId, topicId, orderBy = {}) {
+    const { field = 'updatedAt', direction = 'desc' } = orderBy;
 
     let query = Video.query()
       .select('videos.*', 'users.firstName', 'users.lastName', 'users.email')
       .leftJoin('users', 'videos.userId', 'users.id')
-      .where('topicId', topicId);
+      .where('topicId', topicId)
+      .where('projectId', projectId);
 
     if (field && query) {
       query = query.orderBy(field, direction);
